# CuentaService/src/main/resources/openapi/cuenta-api.yaml
openapi: 3.0.3
info:
  title: Account Service API
  description: Microservice for managing bank accounts and transactions
  version: 1.0.0

servers:
  - url: http://localhost:8080/api/v1

tags:
  - name: Accounts
  - name: Transactions
  - name: Reports

paths:

  /api/v1/accounts:
    post:
      tags: [Accounts]
      operationId: createAccount
      summary: Create a new account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountRequest'
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponse'

    get:
      tags: [Accounts]
      operationId: getAccounts
      summary: Get all accounts or filter by clientId
      parameters:
        - name: clientId
          in: query
          required: false
          schema:
            type: integer
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        "200":
          description: List of accounts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountPageResponse'

  /api/v1/accounts/{accountNumber}:
    get:
      tags: [Accounts]
      operationId: getAccountById
      summary: Get account by account number
      parameters:
        - name: accountNumber
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Account found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponse'

    put:
      tags: [Accounts]
      operationId: updateAccounts
      summary: Update account
      parameters:
        - name: accountNumber
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountRequest'
      responses:
        "200":
          description: Account updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponse'


  /api/v1/transactions:
    post:
      tags: [Transactions]
      operationId: createTransaction
      summary: Create a transaction (deposit or withdrawal)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionRequest'
      responses:
        "201":
          description: Transaction created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionResponse'

  /api/v1/transactions/account/{accountNumber}:
    get:
      tags: [Transactions]
      operationId: getTransactionsByAccountId
      summary: Get transactions by account
      parameters:
        - name: accountNumber
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        "200":
          description: List of transactions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionPageResponse'

  /api/v1/reports:
    get:
      tags: [ Reports ]
      operationId: getReport
      summary: Generate account statement report
      description: >
        Returns the account statement for a client within a date range,
        including accounts, balances and transaction details.

      parameters:
        - name: clientId
          in: query
          required: true
          description: Client identifier
          schema:
            type: integer
            format: int64

        - name: fromDate
          in: query
          required: true
          description: Start date (inclusive)
          schema:
            type: string
            format: date
            example: "2022-02-01"

        - name: toDate
          in: query
          required: true
          description: End date (inclusive)
          schema:
            type: string
            format: date
            example: "2022-02-28"

      responses:
        "200":
          description: Account statement generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountStatementReportResponse'

        "400":
          description: Invalid request

        "404":
          description: Client not found

components:
  parameters:
    Page:
      name: page
      in: query
      required: false
      description: Page number (0-based)
      schema:
        type: integer
        minimum: 0
        default: 0

    Size:
      name: size
      in: query
      required: false
      description: Page size
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20

  schemas:

    AccountRequest:
      type: object
      required:
        - accountNumber
        - accountType
        - initialBalance
        - clientId
      properties:
        accountNumber:
          type: string
          example: "478758"
        accountType:
          type: string
          enum: [SAVINGS, CHECKING]
        initialBalance:
          type: number
          format: double
          example: 2000
        status:
          type: boolean
          example: true
        clientId:
          type: integer
          example: 1

    AccountResponse:
      type: object
      properties:
        id:
          type: integer
        accountNumber:
          type: string
        accountType:
          type: string
        initialBalance:
          type: number
        status:
          type: boolean
        clientId:
          type: integer

    AccountPageResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/AccountResponse'

        totalElements:
          type: integer
          format: int64

        totalPages:
          type: integer

        size:
          type: integer

        number:
          type: integer

        first:
          type: boolean

        last:
          type: boolean

        empty:
          type: boolean

    TransactionRequest:
      type: object
      required:
        - accountId
        - transactionType
        - amount
      properties:
        accountId:
          type: integer
        transactionType:
          type: string
          description: DEPOSIT or WITHDRAWAL
        amount:
          type: number
          example: -500
        description:
          type: string
          example: "Withdrawal of 500"

    TransactionResponse:
      type: object
      properties:
        id:
          type: integer
        accountId:
          type: integer
        date:
          type: string
          format: date-time
        transactionType:
          type: string
        amount:
          type: number
        initialBalance:
          type: number
        availableBalance:
          type: number
        description:
          type: string

    TransactionPageResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/TransactionResponse'

        totalElements:
          type: integer
          format: int64

        totalPages:
          type: integer

        size:
          type: integer

        number:
          type: integer

        first:
          type: boolean

        last:
          type: boolean

        empty:
          type: boolean

    ErrorResponse:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        timestamp:
          type: string
          format: date-time

    AccountStatementReportResponse:
      type: object

      properties:

        clientId:
          type: integer
          format: int64
          example: 2

        clientName:
          type: string
          example: "Marianela Montalvo"

        identification:
          type: string
          example: "5678"

        fromDate:
          type: string
          format: date

        toDate:
          type: string
          format: date

        reportGeneratedAt:
          type: string
          format: date-time

        accounts:
          type: array
          items:
            $ref: '#/components/schemas/AccountStatementAccountResponse'

    AccountStatementAccountResponse:
      type: object

      properties:

        accountId:
          type: integer
          format: int64

        accountNumber:
          type: string

        accountType:
          type: string

        status:
          type: boolean

        openingBalance:
          type: number
          format: double
          example: 100

        closingBalance:
          type: number
          format: double
          example: 700

        transactions:
          type: array
          items:
            $ref: '#/components/schemas/AccountStatementTransactionResponse'

    AccountStatementTransactionResponse:
      type: object

      properties:

        transactionId:
          type: integer
          format: int64

        date:
          type: string
          format: date-time

        transactionType:
          type: string
          example: DEPOSIT

        amount:
          type: number
          format: double
          example: 600

        initialBalance:
          type: number
          format: double
          example: 100

        availableBalance:
          type: number
          format: double
          example: 700